feat(headless-styles): move select option into select folder
