--- packages/headless-styles/src/components/SelectOption/shared.ts
@@ -1,13 +1,0 @@
-import type { SelectOptionOptions } from './types'
-
-export function getDefaultSelectOptionOptions(options?: SelectOptionOptions) {
-  return {
-    value: options?.value ?? '',
-  }
-}
-
-export function createSelectOptionProps(options: SelectOptionOptions) {
-  return {
-    value: options.value,
-  }
-}
