--- packages/headless-styles/src/components/SelectOption/types.ts
@@ -1,3 +1,0 @@
-export interface SelectOptionOptions {
-  value?: string
-}
