+++ packages/headless-styles/src/index.ts
@@ -64,8 +64,8 @@ export { getJSRadioProps } from './components/Radio/radioJS'
 export { getSelectProps } from './components/Select/selectCSS'
 export { getJSSelectProps } from './components/Select/selectJS'
 
-export { getSelectOptionProps } from './components/SelectOption/selectOptionCSS'
-export { getJSSelectOptionProps } from './components/SelectOption/selectOptionJS'
+export { getSelectOptionProps } from './components/Select/selectOptionCSS'
+export { getJSSelectOptionProps } from './components/Select/selectOptionJS'
 
 export { getSkeletonProps } from './components/Skeleton/skeletonCSS'
 export { getJSSkeletonProps } from './components/Skeleton/skeletonJS'
